refector: 오더 폼필드 단순화에 따른 오더명, 코멘트 파싱 제거
